A makefile and cat
