+++ scripts/regen_example.sh
@@ -6,7 +6,7 @@ temp=$(mktemp -q)
 # shellcheck disable=SC2001
 EXPECTED_FILENAME=$(echo "$1" | sed s/expected/actual/)
 
-bundle exec ruby src/rubyfmt.rb "$1" "$EXPECTED_FILENAME" > "$temp" && mv "$temp" "$EXPECTED_FILENAME"
+bundle exec ruby build/rubyfmt.rb "$1" "$EXPECTED_FILENAME" > "$temp" && mv "$temp" "$EXPECTED_FILENAME"
 git diff "$EXPECTED_FILENAME"
 echo "do you want to commit [Y/n]"
 read -rn1 ans
