+++ scripts/test_fixtures.sh
@@ -5,7 +5,7 @@ test_folder() {
     find "$1" -name "*_expected.rb" -maxdepth 1 | while read -r file
     do
         # shellcheck disable=SC2001
-        time ruby --disable=gems src/rubyfmt.rb "$(echo "$file" | sed s/expected/actual/)" > /tmp/out.rb
+        time ruby --disable=gems build/rubyfmt.rb "$(echo "$file" | sed s/expected/actual/)" > /tmp/out.rb
 
         if ! diff -u /tmp/out.rb "$file"
         then
@@ -13,7 +13,7 @@ test_folder() {
             exit 1
         fi
 
-        time ruby --disable=gems src/rubyfmt.rb "$file" > /tmp/out.rb
+        time ruby --disable=gems build/rubyfmt.rb "$file" > /tmp/out.rb
 
         if ! diff -u /tmp/out.rb "$file"
         then
