+++ scripts/test_repo.sh
@@ -11,7 +11,7 @@ repo_path="$1"
 for file in $(git --work-tree="$repo_path" --git-dir="$repo_path/.git" ls-files | grep '\.rb$')
 do
   full_path="$repo_path/$file"
-  errors=$(ruby --disable=gems src/rubyfmt.rb "$full_path" 2>&1)
+  errors=$(ruby --disable=gems build/rubyfmt.rb "$full_path" 2>&1)
   # shellcheck disable=SC2181
   if [[ "$?" -ne 0 ]]
   then
