+++ src/format.rb
@@ -1,492 +1,3 @@
-#!/usr/bin/env ruby
-require "ripper"
-require "stringio"
-require "pp"
-LineMetadata = Struct.new(:comment_blocks)
-
-class HardNewLine
-  def to_s
-    "\n"
-  end
-end
-
-class BreakableEntry
-  def initialize
-    @parts = []
-  end
-
-  def <<(item)
-    @parts << item
-  end
-
-  def to_s
-    @parts.join("")
-  end
-end
-
-class Line
-  attr_accessor :parts
-  def initialize(parts)
-    @comments = []
-    @breakable_entry_stack = []
-    @parts = parts
-  end
-
-  def breakable_entry(&blk)
-    @breakable_entry_stack << BreakableEntry.new
-    blk.call
-    @parts << @breakable_entry_stack.pop
-  end
-
-  def push_comment(comment)
-    @comments << comment
-  end
-
-  def has_comment?
-    !@comments.empty?
-  end
-
-  def <<(item)
-    if be = @breakable_entry_stack.last
-      be << item
-    else
-      @parts << item
-    end
-  end
-
-  def string_length
-    @parts.join("").length
-  end
-
-  def empty?
-    @parts.empty?
-  end
-
-  def to_s
-    build = @parts.join("")
-
-    unless @comments.empty?
-      build = "#{@comments.join("\n")}\n#{build}"
-    end
-
-    build
-  end
-
-  def strip_trailing_newlines
-    while ends_with_newline?
-      @parts.pop
-    end
-  end
-
-  def remove_redundant_indents
-    @parts.shift if @parts[0] == ""
-  end
-
-  def ends_with_newline?
-    HardNewLine === @parts.last
-  end
-
-  def is_only_a_newline?
-    @parts.length == 1 && HardNewLine === @parts[0]
-  end
-
-  def contains_end?
-    @parts.any? { |x| x == :end }
-  end
-
-  def contains_def?
-    @parts.any? { |x| x == :def }
-  end
-
-  def contains_do?
-    @parts.any? { |x| x == :do }
-  end
-
-  def contains_if?
-    @parts.any? { |x| x == :if }
-  end
-
-  def contains_else?
-    @parts.any? { |x| x == :else }
-  end
-
-  def contains_unless?
-    @parts.any? { |x| x == :unless }
-  end
-
-  def declares_private?
-    @parts.any? { |x| x == "private" } && @parts.length == 3
-  end
-
-  def declares_require?
-    @parts.any? { |x| x == "require" } && @parts.none? { |x| x == "}" }
-  end
-
-  def declares_class_or_module?
-    @parts.any? { |x| x == :class || x == :module }
-  end
-
-  def contains_while?
-    @parts.any? { |x| x == :while }
-  end
-
-  def surpresses_blankline?
-    contains_def? || contains_do? || contains_while? || contains_if? || contains_else?
-  end
-end
-
-def want_blankline?(line, next_line)
-  return unless next_line
-  return true if line.contains_end? && !next_line.contains_end?
-  return true if next_line.contains_do? && !line.surpresses_blankline?
-  return true if (next_line.contains_if? || next_line.contains_unless?) && !line.surpresses_blankline?
-  return true if line.declares_private?
-  return true if line.declares_require? && !next_line.declares_require?
-  return true if !line.declares_class_or_module? && next_line.has_comment?
-  return true if !line.declares_class_or_module? && next_line.declares_class_or_module?
-end
-
-class ParserState
-  attr_accessor :depth_stack, :start_of_line, :line, :string_concat_position, :surpress_one_paren
-  attr_reader :heredoc_strings
-  attr_reader :result
-  attr_reader :current_orig_line_number
-  attr_accessor :render_queue
-  attr_reader :comments_hash
-  attr_reader :depth_stack
-  def initialize(result, line_metadata)
-    @surpress_comments_stack = [false]
-    @surpress_one_paren = false
-    @result = result
-    @depth_stack = [0]
-    @start_of_line = [true]
-    @render_queue = []
-    @line = Line.new([])
-    @current_orig_line_number = 0
-    @comments_hash = line_metadata.comment_blocks
-    @conditional_indent = [0]
-    @heredoc_strings = []
-    @string_concat_position = []
-  end
-
-  def self.with_depth_stack(output, from:)
-    i = new(output, LineMetadata.new({}))
-    i.depth_stack = from.depth_stack.dup
-    i
-  end
-
-  def breakable_entry(&blk)
-    @line.breakable_entry(&blk)
-  end
-
-  def with_surpress_comments(value, &blk)
-    @surpress_comments_stack << value
-    blk.call
-    @surpress_comments_stack.pop
-  end
-
-  def push_heredoc_content(symbol, indent, inner_string_components)
-    buf = StringIO.new
-    next_ps = ParserState.new(buf, LineMetadata.new({}))
-    next_ps.depth_stack = depth_stack.dup
-    format_inner_string(next_ps, inner_string_components, :heredoc)
-
-    next_ps.emit_newline
-    next_ps.write
-    buf.rewind
-
-    # buf gets an extra newline on the end, trim it
-    @heredoc_strings << [symbol, indent, buf.read[0...-1]]
-    next_ps.heredoc_strings.each do |s|
-      @heredoc_strings << s
-    end
-  end
-
-  def with_start_of_line(value, &blk)
-    start_of_line << value
-    blk.call
-    start_of_line.pop
-  end
-
-  def start_string_concat
-    push_conditional_indent(:string) if @string_concat_position.empty?
-    @string_concat_position << Object.new
-  end
-
-  def end_string_concat
-    @string_concat_position.pop
-    pop_conditional_indent if @string_concat_position.empty?
-  end
-
-  def on_line(line_number, skip=false)
-    if line_number != @current_orig_line_number
-      @arrays_on_line = -1
-    end
-
-    while !comments_hash.empty? && comments_hash.keys.sort.first < line_number
-      key = comments_hash.keys.sort.first
-      comment = comments_hash.delete(key)
-      @line.push_comment(comment) unless @surpress_comments_stack.last
-    end
-
-    @current_orig_line_number = line_number
-  end
-
-  def write
-    on_line(100000000000000)
-    clear_empty_trailing_lines
-
-    lines = render_queue
-    clear_double_spaces(lines)
-    add_valid_blanklines(lines)
-
-    lines.each do |line|
-      line.remove_redundant_indents
-    end
-
-    ensure_file_ends_with_exactly_one_newline(lines)
-
-    result.write("\n")
-    result.flush
-  end
-
-  def emit_while
-    line << :while
-  end
-
-  def emit_for
-    line << :for
-  end
-
-  def emit_in
-    line << :in
-  end
-
-  def emit_indent
-    spaces = (@conditional_indent.last) + (2 * @depth_stack.last)
-    line << " " * spaces
-  end
-
-  def emit_slash
-    line << "\\"
-  end
-
-  def emit_else
-    line << :else
-  end
-
-  def emit_elsif
-    line << "elsif"
-  end
-
-  def push_conditional_indent(type)
-    if line.empty?
-      @conditional_indent << 2*@depth_stack.last
-    else
-      if type == :conditional
-        @conditional_indent << 2*@depth_stack.last
-      elsif type == :string
-        @conditional_indent << line.string_length
-      end
-    end
-
-    @depth_stack << 0
-  end
-
-  def pop_conditional_indent
-    @conditional_indent.pop
-    @depth_stack.pop
-  end
-
-  def emit_comma_space
-    line << ", "
-  end
-
-  def emit_return
-    line << :return
-  end
-
-
-  def ensure_file_ends_with_exactly_one_newline(lines)
-    lines.each_with_index do |line, i|
-      if i == lines.length-1
-        line.strip_trailing_newlines
-      end
-
-      result.write(line)
-    end
-  end
-
-  def add_valid_blanklines(lines)
-    line = lines.first
-    next_index = 1
-    while next_index < lines.length
-      if want_blankline?(line, lines[next_index])
-        lines.insert(next_index, Line.new(["\n"]))
-        next_index += 1
-      end
-
-      line = lines[next_index]
-      next_index += 1
-    end
-  end
-
-  def clear_double_spaces(lines)
-    line = lines.first
-    next_index = 1
-    while next_index < lines.length
-      while line.ends_with_newline? && lines[next_index] && lines[next_index].is_only_a_newline?
-        lines.delete_at(next_index)
-      end
-
-      line = lines[next_index]
-      next_index += 1
-    end
-  end
-
-  def clear_empty_trailing_lines
-    while render_queue.last == []
-      render_queue.pop
-    end
-
-    while render_queue.last.is_only_a_newline?
-      render_queue.pop
-    end
-  end
-
-  def emit_def(def_name)
-    line << :def
-    line << " #{def_name}"
-  end
-
-  def emit_params_list(params_list)
-  end
-
-  def emit_binary(symbol)
-    line << " #{symbol} "
-  end
-
-  def emit_end
-    emit_newline
-    emit_indent if start_of_line.last
-    line << :end
-  end
-
-  def emit_space
-    line << " "
-  end
-
-  def emit_do
-    line << :do
-  end
-
-  def emit_newline
-    line << HardNewLine.new
-    render_queue << line
-    self.line = Line.new([])
-    render_heredocs
-  end
-
-  def emit_dot
-    line << "."
-  end
-
-  def emit_lonely_operator
-    line << "&."
-  end
-
-  def emit_ident(ident)
-    line << ident
-  end
-
-  def emit_op(op)
-    line << op
-  end
-
-  def emit_int(int)
-    line << int
-  end
-
-  def emit_var_ref(ref)
-    line << ref
-  end
-
-  def emit_open_paren
-    line << "("
-  end
-
-  def emit_close_paren
-    line << ")"
-  end
-
-  def new_block(&blk)
-    depth_stack[-1] += 1
-    with_start_of_line(true, &blk)
-    depth_stack[-1] -= 1
-  end
-
-  def dedent(&blk)
-    depth_stack[-1] -= 1
-    with_start_of_line(true, &blk)
-    depth_stack[-1] += 1
-  end
-
-  def emit_open_block_arg_list
-    line << "|"
-  end
-
-  def emit_close_block_arg_list
-    line << "|"
-  end
-
-  def emit_double_quote
-    line << "\""
-  end
-
-  def emit_module_keyword
-    line << :module
-  end
-
-  def emit_class_keyword
-    line << :class
-  end
-
-  def emit_const(const)
-    line << const
-  end
-
-  def emit_double_colon
-    line << "::"
-  end
-
-  def emit_symbol(symbol)
-    line << ":#{symbol}"
-  end
-
-  def render_heredocs(skip=false)
-    while !heredoc_strings.empty?
-      symbol, indent, string = heredoc_strings.pop
-      unless render_queue[-1] && render_queue[-1].ends_with_newline?
-        line << "\n"
-      end
-
-      if string.end_with?("\n")
-        string = string[0...-1]
-      end
-
-      line << string
-      emit_newline
-      if indent
-        emit_indent
-      end
-      emit_ident(symbol.to_s.gsub("'", ""))
-      if !skip
-        emit_newline
-      end
-    end
-  end
-end
-
 def format_block_params_list(ps, params_list)
   ps.emit_open_block_arg_list
   ps.emit_params_list(params_list)
@@ -2639,164 +2150,3 @@ def format_program(line_metadata, sexp, result)
   ps.write
 end
 
-def extract_line_metadata(file_data)
-  comment_blocks = {}
-
-  file_data.split("\n").each_with_index do |line, index|
-    comment_blocks[index] = line if /^ *#/ === line
-  end
-
-  LineMetadata.new(comment_blocks)
-end
-
-class Parser < Ripper::SexpBuilderPP
-  ARRAY_SYMBOLS = {
-    qsymbols: '%i',
-    qwords: '%w',
-    symbols: '%I',
-    words: '%W'
-  }.freeze
-
-  def self.is_percent_array?(rest)
-    return false if rest.nil?
-    return false if rest[0].nil?
-    ARRAY_SYMBOLS.include?(rest[0][0])
-  end
-
-  def self.percent_symbol_for(rest)
-    ARRAY_SYMBOLS[rest[0][0]]
-  end
-
-  def initialize(file_data)
-    super(file_data)
-    @file_lines = file_data.split("\n")
-    # heredoc stack is the stack of identified heredocs
-    @heredoc_stack = []
-
-    # next_heredoc_stack is the type identifiers of the next heredocs, that
-    # we haven't emitted yet
-    @next_heredoc_stack = []
-    @heredoc_regex = /(<<[-~]?)(.*$)/
-    @next_comment_delete = []
-    @comments_delete = []
-    @regexp_stack = []
-    @string_stack = []
-  end
-
-  attr_reader :comments_delete
-
-  private
-
-  ARRAY_SYMBOLS.each do |event, symbol|
-    define_method(:"on_#{event}_new") do
-      [event, [], [lineno, column]]
-    end
-
-    define_method(:"on_#{event}_add") do |parts, part|
-      parts.tap do |node|
-        node[1] << part
-      end
-    end
-  end
-
-  def on_heredoc_beg(*args, &blk)
-    heredoc_parts = @heredoc_regex.match(args[0]).captures
-    raise "bad heredoc" unless heredoc_parts.select { |x| x != nil }.count == 2
-    @next_heredoc_stack.push(heredoc_parts)
-    @next_comment_delete.push(lineno)
-    super
-  end
-
-  def on_heredoc_end(*args, &blk)
-    @heredoc_stack.push(@next_heredoc_stack.pop)
-    start_com = @next_comment_delete.pop
-    end_com = lineno
-    @comments_delete.push([start_com, end_com])
-    super
-  end
-
-  def on_string_literal(*args, &blk)
-    if @heredoc_stack.last
-      heredoc_parts = @heredoc_stack.pop
-      args.insert(0, [:heredoc_string_literal, heredoc_parts])
-    else
-      end_delim = @string_stack.pop
-      start_delim = @string_stack.pop
-
-      if start_delim != "\""
-        reject_embexpr = start_delim == "'" || start_delim.start_with?("%q")
-
-        (args[0][1..-1] || []).each do |part|
-          next if part.nil?
-          case part[0]
-          when :@tstring_content
-            part[1] = eval("#{start_delim}#{part[1]}#{end_delim}").inspect[1..-2]
-          when :string_embexpr, :string_dvar
-            if reject_embexpr
-              raise "got #{part[0]} in a #{start_delim}...#{end_delim} string"
-            end
-          else
-            raise "got #{part[0]} in a #{start_delim}...#{end_delim} string"
-          end
-        end
-      end
-    end
-    super
-  end
-
-  def on_lambda(*args, &blk)
-    terminator = @file_lines[lineno-1]
-    if terminator.include?("}")
-      args.insert(1, :curly)
-    else
-      args.insert(1, :do)
-    end
-
-    super
-  end
-
-  def on_tstring_beg(*args, &blk)
-    @string_stack << args[0]
-    super
-  end
-
-  def on_tstring_end(*args, &blk)
-    @string_stack << args[0]
-    super
-  end
-
-  def on_regexp_beg(re_part)
-    @regexp_stack << re_part
-  end
-
-  def on_regexp_literal(*args)
-    args[1] << @regexp_stack.pop
-    super(*args)
-  end
-end
-
-def main
-  file_data = ARGF.read
-  file_data = file_data.gsub("\r\n", "\n")
-
-  line_metadata = extract_line_metadata(file_data)
-
-  parser = Parser.new(file_data)
-  sexp = parser.parse
-  if ENV["RUBYFMT_DEBUG"] == "2"
-    require 'pry'; binding.pry
-  end
-  if parser.error?
-    if ENV["RUBYFMT_DEBUG"] == "2"
-      require 'pry'; binding.pry
-    end
-    raise parser.error
-  end
-
-  parser.comments_delete.each do |(start, last)|
-    line_metadata.comment_blocks.reject! { |k, v| k >= start && k <= last }
-  end
-  format_program(line_metadata, sexp, $stdout)
-end
-
-main if __FILE__ == $0
