+++ src/line_metadata.rb
@@ -1,0 +1,1 @@
+LineMetadata = Struct.new(:comment_blocks)
