+++ src/requires.rb
@@ -1,0 +1,3 @@
+#!/usr/bin/env ruby
+require "ripper"
+require "stringio"
